% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliableChangeIndex.R
\name{RCI}
\alias{RCI}
\title{Reliable Change Index (RCI) value}
\usage{
RCI(x1, x2, s1, s2, rxy, PE)
}
\arguments{
\item{x1}{baseline score (pretest).}

\item{x2}{retest score.}

\item{s1}{baseline standard deviation.}

\item{s2}{retest standard deviation.}

\item{rxy}{test retest reliability of measure.}

\item{PE}{Control group difference score or the practice effect (Mean score control after – Mean score control before).}
}
\value{
The RCI values based on the input value
}
\description{
This function calculates Reliable Change Index (RCI) value
}
\examples{
\dontrun{
RCI (104,96,13.8,15,0.95,4.3)
}

}
